requirements.txt
streamlit>=1.45,<2
python-docx>=1.1,<2
